test: add CP zigzag tests
